requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pymongo>=4.6.0
pydantic>=2.0.0
python-dotenv>=1.0.0
crewai>=0.28.0
crewai-tools>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.10
pytest>=7.4.0
pyyaml>=6.0